+++ src/aioway/spaces/__init__.py
@@ -2,9 +2,4 @@
 
 from .casts import *
 from .data import *
-from .dims import *
-from .gym import *
-from .media import *
 from .spaces import *
-from .tdicts import *
-from .tensors import *
