+++ src/aioway/spaces/data/_bases.py
