+++ src/aioway/spaces/data/attrs.py
