+++ src/aioway/spaces/data/data.py
@@ -7,7 +7,7 @@ import abc
 from aioway._api import public_api
 from aioway._utils import torch_fake_mode
 
-from .spaces import Space, space_dcls
+from ..spaces import Space, space_dcls
 
 __all__ = ["DataSpace"]
