+++ src/aioway/spaces/data/devices.py
