+++ src/aioway/spaces/data/dims.py
@@ -9,8 +9,9 @@ import typing
 
 import torch
 
-from .spaces import space_dcls
-from .tensors import Attr, TensorSpace
+from ..spaces import space_dcls
+from .attrs import Attr
+from .tensors import TensorSpace
 
 __all__ = ["DimSpace", "DimInfo"]
