+++ src/aioway/spaces/data/shapes.py
