+++ src/aioway/spaces/data/tdicts.py
@@ -14,9 +14,10 @@ import tensordict as td
 from aioway._api import public_api
 from aioway._utils import is_real_tensor, torch_fake_mode
 
+from ..spaces import space_dcls
+from .attrs import Attr, AttrLike
 from .data import DataSpace
-from .spaces import space_dcls
-from .tensors import Attr, AttrLike, DType
+from .dtypes import DType
 
 __all__ = ["Schema", "TdictSpace", "TensorClassSpace"]
